#1 chore: 디렉토리명 변경
