+++ src/main/java/ollie/wecare/common/base/BaseException.java
@@ -1,4 +1,4 @@
-package ollie.wecare.common.Base;
+package ollie.wecare.common.base;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
