+++ src/main/java/ollie/wecare/common/base/BaseResponse.java
@@ -1,4 +1,4 @@
-package ollie.wecare.common.Base;
+package ollie.wecare.common.base;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import static ollie.wecare.common.Base.BaseResponseStatus.SUCCESS;
+import static ollie.wecare.common.base.BaseResponseStatus.SUCCESS;
 
 @Getter
 @AllArgsConstructor
@@ -15,8 +15,8 @@ public class BaseResponse<T> {
 
     @JsonProperty("isSuccess")
     private final Boolean isSuccess;
+
     private final String message;
-    private final int code;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private T result;
@@ -24,13 +24,11 @@ public class BaseResponse<T> {
     public BaseResponse(T result) {
         this.isSuccess = SUCCESS.isSuccess();
         this.message = SUCCESS.getMessage();
-        this.code = SUCCESS.getCode();
         this.result = result;
     }
 
     public BaseResponse(BaseResponseStatus status) {
         this.isSuccess = status.isSuccess();
         this.message = status.getMessage();
-        this.code = status.getCode();
     }
 }
